+++ generate/testdata/snapshots/TestGenerate-InputEnum.graphql-InputEnum.graphql.go
@@ -54,7 +54,7 @@ type __InputEnumQueryInput struct {
 // GetRole returns __InputEnumQueryInput.Role, and is useful for accessing the field via an interface.
 func (v *__InputEnumQueryInput) GetRole() Role { return v.Role }
 
-// The query or mutation executed by InputEnumQuery.
+// The query executed by InputEnumQuery.
 const InputEnumQuery_Operation = `
 query InputEnumQuery ($role: Role!) {
 	usersWithRole(role: $role) {
@@ -66,7 +66,7 @@ query InputEnumQuery ($role: Role!) {
 func InputEnumQuery(
 	client_ graphql.Client,
 	role Role,
-) (*InputEnumQueryResponse, error) {
+) (data_ *InputEnumQueryResponse, err_ error) {
 	req_ := &graphql.Request{
 		OpName: "InputEnumQuery",
 		Query:  InputEnumQuery_Operation,
@@ -74,10 +74,9 @@ func InputEnumQuery(
 			Role: role,
 		},
 	}
-	var err_ error
 
-	var data_ InputEnumQueryResponse
-	resp_ := &graphql.Response{Data: &data_}
+	data_ = &InputEnumQueryResponse{}
+	resp_ := &graphql.Response{Data: data_}
 
 	err_ = client_.MakeRequest(
 		nil,
@@ -85,6 +84,6 @@ func InputEnumQuery(
 		resp_,
 	)
 
-	return &data_, err_
+	return data_, err_
 }