+++ generate/testdata/snapshots/TestGenerateWithConfig-ExportOperations-testdata-queries-generated.go
@@ -34,7 +34,7 @@ type SimpleQueryUser struct {
 // GetId returns SimpleQueryUser.Id, and is useful for accessing the field via an interface.
 func (v *SimpleQueryUser) GetId() string { return v.Id }
 
-// The query or mutation executed by SimpleQuery.
+// The query executed by SimpleQuery.
 const SimpleQuery_Operation = `
 query SimpleQuery {
 	user {
@@ -46,15 +46,14 @@ query SimpleQuery {
 func SimpleQuery(
 	ctx_ context.Context,
 	client_ graphql.Client,
-) (*SimpleQueryResponse, error) {
+) (data_ *SimpleQueryResponse, err_ error) {
 	req_ := &graphql.Request{
 		OpName: "SimpleQuery",
 		Query:  SimpleQuery_Operation,
 	}
-	var err_ error
 
-	var data_ SimpleQueryResponse
-	resp_ := &graphql.Response{Data: &data_}
+	data_ = &SimpleQueryResponse{}
+	resp_ := &graphql.Response{Data: data_}
 
 	err_ = client_.MakeRequest(
 		ctx_,
@@ -62,6 +61,6 @@ func SimpleQuery(
 		resp_,
 	)
 
-	return &data_, err_
+	return data_, err_
 }