+++ generate/testdata/snapshots/TestGenerateWithConfig-StructReferences-testdata-queries-generated.go
@@ -226,7 +226,7 @@ type __InputObjectQueryInput struct {
 // GetQuery returns __InputObjectQueryInput.Query, and is useful for accessing the field via an interface.
 func (v *__InputObjectQueryInput) GetQuery() *UserQueryInput { return v.Query }
 
-// The query or mutation executed by InputObjectQuery.
+// The query executed by InputObjectQuery.
 const InputObjectQuery_Operation = `
 query InputObjectQuery ($query: UserQueryInput) {
 	user(query: $query) {
@@ -239,7 +239,7 @@ func InputObjectQuery(
 	ctx_ context.Context,
 	client_ graphql.Client,
 	query *UserQueryInput,
-) (*InputObjectQueryResponse, error) {
+) (data_ *InputObjectQueryResponse, err_ error) {
 	req_ := &graphql.Request{
 		OpName: "InputObjectQuery",
 		Query:  InputObjectQuery_Operation,
@@ -247,10 +247,9 @@ func InputObjectQuery(
 			Query: query,
 		},
 	}
-	var err_ error
 
-	var data_ InputObjectQueryResponse
-	resp_ := &graphql.Response{Data: &data_}
+	data_ = &InputObjectQueryResponse{}
+	resp_ := &graphql.Response{Data: data_}
 
 	err_ = client_.MakeRequest(
 		ctx_,
@@ -258,10 +257,10 @@ func InputObjectQuery(
 		resp_,
 	)
 
-	return &data_, err_
+	return data_, err_
 }
 
-// The query or mutation executed by QueryWithStructs.
+// The query executed by QueryWithStructs.
 const QueryWithStructs_Operation = `
 query QueryWithStructs {
 	user {
@@ -276,15 +275,14 @@ query QueryWithStructs {
 func QueryWithStructs(
 	ctx_ context.Context,
 	client_ graphql.Client,
-) (*QueryWithStructsResponse, error) {
+) (data_ *QueryWithStructsResponse, err_ error) {
 	req_ := &graphql.Request{
 		OpName: "QueryWithStructs",
 		Query:  QueryWithStructs_Operation,
 	}
-	var err_ error
 
-	var data_ QueryWithStructsResponse
-	resp_ := &graphql.Response{Data: &data_}
+	data_ = &QueryWithStructsResponse{}
+	resp_ := &graphql.Response{Data: data_}
 
 	err_ = client_.MakeRequest(
 		ctx_,
@@ -292,6 +290,6 @@ func QueryWithStructs(
 		resp_,
 	)
 
-	return &data_, err_
+	return data_, err_
 }